MI5-44 Kreiran MockTransactionsLoader
MockTransactionsLoader služi kako bi vratio listu mock transakcija. MockTransactionsLoader se koristi u TransactionsCandidatesViewu sve dok se ne postavi logika za dohvat transakcija kandidata.
